+++ packages/cache/src/options.ts
@@ -1,0 +1,17 @@
+/**
+ * Options to control cache upload
+ */
+export interface UploadOptions {
+  /**
+   * Number of parallel cache upload
+   *
+   * @default 4
+   */
+  uploadConcurrency?: number
+  /**
+   * Maximum chunk size in bytes for cache upload
+   *
+   * @default 32MB
+   */
+  uploadChunkSize?: number
+}